feat: implement AI provider architecture with Gemini and Deepseek support
