+++ .env.example
@@ -1,0 +1,34 @@
+# AI Provider API Keys (Required)
+GEMINI_API_KEY=your_gemini_api_key_here
+DEEPSEEK_API_KEY=your_deepseek_api_key_here
+OPENAI_API_KEY=your_openai_api_key_here    # Optional
+ANTHROPIC_API_KEY=your_claude_api_key_here  # Optional
+
+# Provider Configuration
+ENABLED_AI_PROVIDERS=gemini,deepseek
+AI_PROVIDER_FALLBACK_CHAIN=gemini,deepseek
+
+# Gemini Settings
+GEMINI_MODEL=gemini-2.0-flash-001
+GEMINI_TEMPERATURE=0.8
+GEMINI_TOP_P=0.95
+GEMINI_MAX_TOKENS=8192
+
+# DeepSeek Settings
+DEEPSEEK_TEMPERATURE=0.7
+
+# ChatGPT Settings (Optional)
+OPENAI_MODEL=gpt-4
+OPENAI_TEMPERATURE=0.7
+OPENAI_MAX_TOKENS=4000
+OPENAI_TOP_P=0.95
+OPENAI_FREQ_PENALTY=0
+OPENAI_PRES_PENALTY=0
+
+# Claude Settings (Optional)
+CLAUDE_MODEL=claude-3-opus-20240229
+CLAUDE_TEMPERATURE=0.7
+CLAUDE_MAX_TOKENS=4000
+
+# GitHub Configuration
+GITHUB_TOKEN=your_github_token_here  # Required for GitHub API access
