+++ review_code.py
@@ -7,170 +7,16 @@ import difflib
 import requests
 import fnmatch
 from unidiff import Hunk, PatchedFile, PatchSet
-from abc import ABC, abstractmethod
+from ai_providers.factory import AIProviderFactory
 
 GITHUB_TOKEN = os.environ["GITHUB_TOKEN"]
 gh = Github(GITHUB_TOKEN)
 
-class AIProvider(ABC):
-    """Abstract base class for AI code review providers."""
-    
-    @abstractmethod
-    def configure(self) -> None:
-        """Configure the AI provider with necessary credentials and settings."""
-        pass
-
-    @abstractmethod
-    def generate_review(self, prompt: str) -> List[Dict[str, Any]]:
-        """Generate code review from the given prompt."""
-        pass
-
-    @abstractmethod
-    def get_name(self) -> str:
-        """Get the name of the AI provider."""
-        pass
-
-class GeminiProvider(AIProvider):
-    """Gemini AI provider implementation."""
-
-    def __init__(self):
-        self.model = None
-        self.generation_config = {
-            "max_output_tokens": 8192,
-            "temperature": 0.8,
-            "top_p": 0.95,
-        }
-
-    def configure(self) -> None:
-        """Configure Gemini with API key and model."""
-        Client.configure(api_key=os.environ.get('GEMINI_API_KEY'))
-        model_name = os.environ.get('GEMINI_MODEL', 'gemini-2.0-flash-001')
-        self.model = Client.GenerativeModel(model_name)
-
-    def generate_review(self, prompt: str) -> List[Dict[str, Any]]:
-        """Generate code review using Gemini AI."""
-        try:
-            response = self.model.generate_content(prompt, generation_config=self.generation_config)
-            
-            response_text = response.text.strip()
-            if response_text.startswith('```json'):
-                response_text = response_text[7:]
-            if response_text.endswith('```'):
-                response_text = response_text[:-3]
-            response_text = response_text.strip()
-
-            try:
-                data = json.loads(response_text)
-                if "reviews" in data and isinstance(data["reviews"], list):
-                    return [
-                        review for review in data["reviews"]
-                        if "lineNumber" in review and "reviewComment" in review
-                    ]
-            except json.JSONDecodeError as e:
-                print(f"Error decoding JSON response: {e}")
-                return []
-        except Exception as e:
-            print(f"Error during Gemini API call: {e}")
-            return []
-        
-        return []
-
-    def get_name(self) -> str:
-        """Get the provider name."""
-        return "Gemini AI"
-
-class DeepseekProvider(AIProvider):
-    """Deepseek AI provider implementation."""
-
-    def __init__(self):
-        self.api_url = "https://api.deepseek.com/v1/chat/completions"
-        self.api_key = None
-        self.model = None
-        self.config = {
-            "temperature": 0.8,
-            "max_tokens": 8192
-        }
-
-    def configure(self) -> None:
-        """Configure Deepseek with API key and model."""
-        self.api_key = os.environ.get('DEEPSEEK_API_KEY')
-        if not self.api_key:
-            raise ValueError("DEEPSEEK_API_KEY environment variable is required")
-        self.model = os.environ.get('DEEPSEEK_MODEL', 'deepseek-chat')
-
-    def generate_review(self, prompt: str) -> List[Dict[str, Any]]:
-        """Generate code review using Deepseek AI."""
-        try:
-            headers = {
-                "Content-Type": "application/json",
-                "Authorization": f"Bearer {self.api_key}"
-            }
-            
-            data = {
-                "model": self.model,
-                "messages": [
-                    {"role": "user", "content": prompt}
-                ],
-                "temperature": self.config["temperature"],
-                "max_tokens": self.config["max_tokens"]
-            }
-
-            response = requests.post(self.api_url, headers=headers, json=data)
-            response.raise_for_status()
-            
-            response_data = response.json()
-            response_text = response_data['choices'][0]['message']['content'].strip()
-
-            if response_text.startswith('```json'):
-                response_text = response_text[7:]
-            if response_text.endswith('```'):
-                response_text = response_text[:-3]
-            response_text = response_text.strip()
-
-            try:
-                data = json.loads(response_text)
-                if "reviews" in data and isinstance(data["reviews"], list):
-                    return [
-                        review for review in data["reviews"]
-                        if "lineNumber" in review and "reviewComment" in review
-                    ]
-            except json.JSONDecodeError as e:
-                print(f"Error decoding JSON response: {e}")
-                return []
-        except Exception as e:
-            print(f"Error during Deepseek API call: {e}")
-            return []
-        
-        return []
-
-    def get_name(self) -> str:
-        """Get the provider name."""
-        return "Deepseek AI"
-
-class AIProviderFactory:
-    """Factory class for creating and managing AI providers."""
-    
-    _providers = {
-        'gemini': GeminiProvider,
-        'deepseek': DeepseekProvider
-    }
-
-    @classmethod
-    def get_provider(cls, provider_name: str) -> AIProvider:
-        """Get an instance of the specified AI provider."""
-        provider_class = cls._providers.get(provider_name.lower())
-        if not provider_class:
-            supported = list(cls._providers.keys())
-            raise ValueError(f"Unsupported AI provider: {provider_name}. Supported providers: {supported}")
-        
-        provider = provider_class()
-        provider.configure()
-        return provider
-
-# Initialize AI provider
-ai_provider = AIProviderFactory.get_provider(os.environ.get('AI_PROVIDER', 'deepseek'))
+# Initialize AI provider from the factory
+provider_name = os.environ.get('AI_PROVIDER', 'deepseek')
+print(f"Creating AI provider: {provider_name}")
+ai_provider = AIProviderFactory.get_provider(provider_name)
 print(f"Initialized AI provider: {ai_provider.get_name()}")
-print(f"AI provider model: {ai_provider.model}")
 
 class PRDetails:
     def __init__(self, owner: str, repo: str, pull_number: int, title: str, description: str):
